requirements.txt
django
djangorestframework
psycopg2
debugpy
drf-writable-nested